Move readJSONOrNull() to utils
